requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
asyncpg==0.28.0
psycopg2-binary==2.9.9